Use 'getpass' to handle the password.
This allows the user to specify the password retrieval method and
avoid leaking their password into the process table and shell
history.

If '-password' is not specified, the user is prompted for the
password on the tty; if the user wants to provide the password
as before, they would pass '-password pass:password'.

See https://github.com/jschauma/getpass for all supported options.

See also:
https://www.netmeister.org/blog/passing-passwords.html